fix: consolidate factorysessionexecution fixtures for pkg-file-count gate
Merge fixture catalog helpers and published-scenario tests to satisfy the
15-file package limit and remove the prohibited source-scanning vocabulary test.
